[2nd Commit] 1. Split Class Matcher 로 로직 변경 2. 테스트코드 변경 3. Pattern 처럼 스레드세이프한 객체의 상수화. 네이밍은 대문자? 소문자?

+++ src/main/java/domain/dto/Operand.java
@@ -2,22 +2,22 @@ package domain.dto;
 
 public class Operand {
 
-    private final double value;
+    private final long value;
 
     public Operand(final String operand) {
-        try{
-            double parsedOperand = Double.parseDouble(operand);
+        try {
+            long parsedOperand = Long.parseLong(operand);
 
             checkNegative(parsedOperand);
 
             this.value = parsedOperand;
-        }catch(NumberFormatException ne){
+        } catch (NumberFormatException ne) {
             throw new IllegalArgumentException(String.format("%s 는 Double 형으로 파싱될 수 없습니다.", operand));
         }
     }
 
     private void checkNegative(double parsedOperand) {
-        if(parsedOperand < 0 ){
+        if (parsedOperand < 0) {
             throw new IllegalArgumentException(String.format("음수인 %s 값은 들어올 수 없습니다", parsedOperand));
         }
     }
